chore: split formatting/regex/collect_modules under the file-size gate
Peel three oversized Rust files into topical sibling modules so they drop
under the 2,000-LOC gate, and remove them from the check_file_size.sh
allowlist. Pure code moves (verified byte-identical to HEAD); no behavior
change.

- builtins/formatting.rs (2220 -> 1794): move the `util.format` /
  `util.formatWithOptions` / `util.inspect` family + their `%j`
  cycle-detection helpers into builtins/formatting/util_format.rs. The
  shared `format_jsvalue` / `format_jsvalue_for_json` value formatter and
  the deep-equal machinery stay in the parent.
- regex.rs (2108 -> 1651): move the `String.prototype.replace`
  substitution machinery (`expand_js_replacement` $-expansion +
  `replace_regex_fn_fancy` fancy-regex callback fallback) into
  regex/replace_expand.rs.
- compile/collect_modules.rs (2032 -> 1788): extract the dynamic-import
  glob expansion into collect_modules/dynamic_glob.rs and relocate the
  inline test module to collect_modules/tests.rs.

Also allowlist perry-dispatch/src/ui_table.rs (2008 LOC) with a rationale:
it is a flat PERRY_UI_TABLE `MethodRow` data array (generated-feel manifest
like entries.rs) that crossed the limit on current main and can't be split
without scattering related rows. This unbreaks the gate, which was already
red on main for that file.

Verified: cargo check (all touched crates) + targeted unit tests pass;
console.log/util.format/util.inspect and regex replace/replaceAll smoke
tests match Node byte-for-byte; check_file_size.sh now exits 0.
