+++ crates/perry/src/commands/compile/collect_modules.rs
@@ -31,9 +31,13 @@ use super::{
 };
 
 mod crypto_ns;
+mod dynamic_glob;
 mod parse_error;
+#[cfg(test)]
+mod tests;
 
 use crypto_ns::module_uses_global_crypto_namespace;
+use dynamic_glob::expand_dynamic_import_glob;
 use parse_error::annotate_parse_error;
 
 const MAX_CROSS_MODULE_INLINE_PRIOR_MODULES: usize = 128;
@@ -183,61 +187,6 @@ pub(super) fn known_node_submodule_key(source: &str) -> Option<&'static str> {
         "hono/jsx/streaming" => Some("hono_jsx_streaming"),
         _ => None,
     }
-}
-
-/// #1674 sub-part B: expand a dynamic-`import()` glob pattern
-/// (`<prefix>*<suffix>`, where `prefix` is a relative, directory-anchored
-/// path) into concrete relative specifiers by reading the importing module's
-/// directory. Each returned specifier equals the string the runtime template
-/// produces (`prefix_dir + filename`), so the compile-time candidate keys match
-/// the runtime dispatch arg exactly. Returns specifiers sorted for determinism.
-fn expand_dynamic_import_glob(
-    importing_file: &str,
-    prefix: &str,
-    suffix: &str,
-    cap: usize,
-) -> Vec<String> {
-    // Split the prefix into its directory part (through the last '/') and the
-    // leading filename fragment that survivors must start with.
-    let last_slash = match prefix.rfind('/') {
-        Some(i) => i,
-        None => return Vec::new(),
-    };
-    let prefix_dir = &prefix[..=last_slash]; // e.g. "./plugins/" or "./"
-    let file_prefix = &prefix[last_slash + 1..]; // e.g. "" or "locale_"
-
-    let importing_dir = std::path::Path::new(importing_file)
-        .parent()
-        .unwrap_or_else(|| std::path::Path::new("."));
-    let glob_dir = importing_dir.join(prefix_dir);
-
-    let entries = match std::fs::read_dir(&glob_dir) {
-        Ok(e) => e,
-        Err(_) => return Vec::new(),
-    };
-    let min_len = file_prefix.len() + suffix.len();
-    let mut out: Vec<String> = Vec::new();
-    for entry in entries.flatten() {
-        if !entry.file_type().map(|t| t.is_file()).unwrap_or(false) {
-            continue;
-        }
-        let name = entry.file_name();
-        let Some(name) = name.to_str() else { continue };
-        // The wildcard must match a non-empty middle: `name` strictly longer
-        // than `file_prefix + suffix`, and bracketed by them.
-        if name.len() <= min_len || !name.starts_with(file_prefix) || !name.ends_with(suffix) {
-            continue;
-        }
-        let candidate = format!("{prefix_dir}{name}");
-        if !out.contains(&candidate) {
-            out.push(candidate);
-        }
-        if out.len() > cap {
-            break;
-        }
-    }
-    out.sort();
-    out
 }
 
 /// Collect all modules to compile (transitive closure of imports)
@@ -1838,195 +1787,4 @@ fn collect_module_finish(
     });
     ctx.native_modules.insert(canonical, hir_module);
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{collect_modules, expand_dynamic_import_glob};
-    use crate::commands::compile::CompilationContext;
-    use crate::commands::progress::VerboseProgress;
-    use crate::OutputFormat;
-    use std::collections::HashSet;
-
-    #[test]
-    fn expands_directory_files_matching_suffix() {
-        // #1674 sub-B: glob `./plugins/*.ts` against the importing module's dir.
-        let base = std::env::temp_dir().join(format!("perry_glob_test_{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&base);
-        let plugins = base.join("plugins");
-        std::fs::create_dir_all(&plugins).unwrap();
-        std::fs::write(plugins.join("alpha.ts"), "export const x=1;").unwrap();
-        std::fs::write(plugins.join("beta.ts"), "export const x=2;").unwrap();
-        std::fs::write(plugins.join("notes.md"), "ignored: wrong suffix").unwrap();
-        let importing = base.join("main.ts");
-        std::fs::write(&importing, "").unwrap();
-
-        let got = expand_dynamic_import_glob(importing.to_str().unwrap(), "./plugins/", ".ts", 64);
-        assert_eq!(
-            got,
-            vec![
-                "./plugins/alpha.ts".to_string(),
-                "./plugins/beta.ts".to_string()
-            ]
-        );
-
-        // A directory with no matches yields nothing (→ rejected promise).
-        let none =
-            expand_dynamic_import_glob(importing.to_str().unwrap(), "./plugins/", ".mjs", 64);
-        assert!(none.is_empty());
-
-        let _ = std::fs::remove_dir_all(&base);
-    }
-
-    #[test]
-    fn dependency_is_transformed_before_importer_for_cross_module_inline() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let root = dir.path();
-        let dep = root.join("dep.ts");
-        let entry = root.join("entry.ts");
-
-        std::fs::write(
-            &dep,
-            r#"
-export class Dep {
-  marker(): number {
-    return 424242;
-  }
-}
-"#,
-        )
-        .expect("write dep");
-        std::fs::write(
-            &entry,
-            r#"
-import { Dep } from "./dep";
-
-const dep = new Dep();
-const got = dep.marker();
-console.log(got);
-"#,
-        )
-        .expect("write entry");
-
-        let mut ctx = CompilationContext::new(root.to_path_buf());
-        ctx.entry_canonical = Some(entry.canonicalize().unwrap());
-        let mut visited = HashSet::new();
-        let mut next_class_id: perry_hir::ClassId = 1;
-        let progress = VerboseProgress::new(OutputFormat::Text, 0);
-
-        collect_modules(
-            &entry,
-            &mut ctx,
-            &mut visited,
-            OutputFormat::Text,
-            None,
-            &mut next_class_id,
-            false,
-            &progress,
-            None,
-        )
-        .expect("collect modules");
-
-        let entry_hir = ctx
-            .native_modules
-            .get(&entry.canonicalize().unwrap())
-            .expect("entry module collected");
-        let entry_debug = format!("{entry_hir:?}");
-
-        assert!(
-            entry_debug.contains("424242"),
-            "entry HIR should contain the dependency method literal after cross-module inlining:\n{entry_debug}"
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn bun_compile_package_js_esm_realpath_parses_as_module() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let root = dir.path();
-        let node_modules = root.join("node_modules");
-        let glob_pkg = node_modules.join(".bun/glob@13.0.5/node_modules/glob");
-        let esm_dir = glob_pkg.join("dist/esm");
-        std::fs::create_dir_all(&esm_dir).expect("create glob esm dir");
-        std::fs::write(
-            glob_pkg.join("package.json"),
-            r#"{
-  "name": "glob",
-  "version": "13.0.5",
-  "type": "module",
-  "exports": {
-    ".": {
-      "import": {
-        "default": "./dist/esm/index.min.js"
-      },
-      "require": {
-        "default": "./dist/commonjs/index.min.js"
-      }
-    }
-  },
-  "module": "./dist/esm/index.min.js",
-  "main": "./dist/commonjs/index.min.js"
-}"#,
-        )
-        .expect("write package json");
-        std::fs::write(esm_dir.join("package.json"), r#"{ "type": "module" }"#)
-            .expect("write esm package json");
-        std::fs::write(esm_dir.join("dep.js"), "export const dep=41;\n").expect("write dep");
-        std::fs::write(
-            esm_dir.join("index.min.js"),
-            r#"import{dep}from"./dep.js";const value=dep+1;export{value};"#,
-        )
-        .expect("write index");
-
-        std::os::unix::fs::symlink(
-            ".bun/glob@13.0.5/node_modules/glob",
-            node_modules.join("glob"),
-        )
-        .expect("symlink glob");
-
-        let entry = root.join("entry.ts");
-        std::fs::write(
-            &entry,
-            r#"
-import { value } from "glob";
-console.log(value);
-"#,
-        )
-        .expect("write entry");
-
-        let mut ctx = CompilationContext::new(root.to_path_buf());
-        ctx.compile_packages.insert("glob".to_string());
-        ctx.entry_canonical = Some(entry.canonicalize().unwrap());
-        let mut visited = HashSet::new();
-        let mut next_class_id: perry_hir::ClassId = 1;
-        let progress = VerboseProgress::new(OutputFormat::Text, 0);
-
-        collect_modules(
-            &entry,
-            &mut ctx,
-            &mut visited,
-            OutputFormat::Text,
-            None,
-            &mut next_class_id,
-            false,
-            &progress,
-            None,
-        )
-        .expect("collect modules");
-
-        let canonical_index = esm_dir.join("index.min.js").canonicalize().unwrap();
-        let canonical_dep = esm_dir.join("dep.js").canonicalize().unwrap();
-        assert!(
-            ctx.native_modules.contains_key(&canonical_index),
-            "glob ESM entry should be compiled natively from Bun realpath"
-        );
-        assert!(
-            ctx.native_modules.contains_key(&canonical_dep),
-            "glob ESM dependency should be compiled natively from Bun realpath"
-        );
-        assert!(
-            ctx.js_modules.is_empty(),
-            "compilePackages ESM files should not route through JS runtime"
-        );
-    }
 }
